MTA-3311: Upgrade PHPUnit to the latest version
- update functional tests
