+++ tests/app/Magento/BlockRender/Test/Block/BlockRenderForm.php
@@ -9,8 +9,7 @@ namespace Magento\BlockRender\Test\Block;
 use Magento\Mtf\Test\Block\TestBlock;
 
 /**
- * Class BlockRender
- * Block for render form
+ * Block for render form.
  */
 class BlockRenderForm extends TestBlock
 {
