+++ tests/app/Magento/BlockRender/Test/Page/Area/TestPage.xml
@@ -8,8 +8,8 @@
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="../../../../../../../etc/pages.xsd">
     <page mca="/" module="Magento_BlockRender" name="TestPage" area="Area">
         <block name="blockRender" class="Magento\BlockRender\Test\Block\BlockRender" locator="body" strategy="tag name">
-            <render name="test" class="Magento\Mtf\Test\Block\TestBlock" locator="#viewport" strategy="css selector"/>
-            <render name="render" class="Magento\BlockRender\Test\Block\BlockRenderForm" locator="#viewport" strategy="css selector"/>
+            <render name="test" class="Magento\Mtf\Test\Block\TestBlock" locator="#viewport" strategy="css selector" />
+            <render name="render" class="Magento\BlockRender\Test\Block\BlockRenderForm" locator="#viewport" strategy="css selector" />
         </block>
     </page>
 </config>
