--- tests/app/Magento/BlockRender/Test/Repository/Test.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" ?>
-<!--
-/**
- * Copyright © 2016 Magento. All rights reserved.
- * See COPYING.txt for license details.
- */
--->
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-            xsi:noNamespaceSchemaLocation="../../../../../../Magento/Mtf/Repository/etc/repository.xsd">
-    <repository class="Magento\Mtf\Test\Repository\Test">
-        <dataset name="some_dummy_configuration">
-            <field path="path/to/dummy/field" scope="dummy" scope_id="1" label="Label value" xsi:type="string">1</field>
-            <field path="path/to/other/dummy/field" scope="dummy" scope_id="1" label="Dummy Label" xsi:type="string">10</field>
-        </dataset>
-    </repository>
-</config>
