--- tests/app/Magento/Mtf/Test/Fixture/BlockRenderMtf.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-/**
- * Copyright © 2016 Magento. All rights reserved.
- * See COPYING.txt for license details.
- */
- -->
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="../../../../../../etc/fixture.xsd">
-    <fixture name="blockRenderMtf"
-             module="Magento_Test"
-             class="Magento\Mtf\Test\Fixture\BlockRenderMtf"
-             extends="\Magento\BlockRender\Test\Fixture\BlockRender"
-            >
-        <field name="strategy" is_required="1" source="Magento\Mtf\Test\Fixture\BlockRenderMtf\Strategy">
-            <default_value xsi:type="string">link text</default_value>
-        </field>
-        <field name="type"/>
-    </fixture>
-</config>
