+++ tests/app/Magento/Mtf/Test/Handler/Test/Ui.php
@@ -10,11 +10,13 @@ use Magento\Mtf\Fixture\FixtureInterface;
 use Magento\Mtf\Handler\Ui as AbstractUi;
 
 /**
- * Class Ui
+ * UI handler for Test fixture.
  */
 class Ui extends AbstractUi implements TestInterface
 {
     /**
+     * Persist Test fixture.
+     *
      * @param FixtureInterface $fixture
      * @return void
      */
