--- tests/app/Magento/Mtf/Test/Repository/BlockRender.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" ?>
-<!--
-/**
- * Copyright © 2016 Magento. All rights reserved.
- * See COPYING.txt for license details.
- */
--->
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-            xsi:noNamespaceSchemaLocation="../../../../../../Magento/Mtf/Repository/etc/repository.xsd">
-    <repository class="Magento\BlockRender\Test\Repository\BlockRender">
-        <dataset name="block_data">
-            <field name="strategy" xsi:type="const">\Magento\Tag\Test\TestCase\TagTest::SEVERITY</field>
-            <field name="type" xsi:type="string">type</field>
-        </dataset>
-    </repository>
-</config>
