+++ tests/app/Magento/Mtf/Test/TestCase/InjectableTest.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+/**
+ * Copyright © 2016 Magento. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+ -->
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="../../../../../../etc/variations.xsd">
+    <testCase name="Magento\Mtf\Test\TestCase\InjectableTest" summary="Test Injectable test case">
+        <variation name="InjectableTestCaseVariation1" method="test1">
+            <data name="fixture/data/locator" xsi:type="string">#hplogo</data>
+            <data name="fixture/data/strategy" xsi:type="string">css selector</data>
+            <constraint name="Magento\Mtf\Test\Constraint\PageOpenSuccess" />
+        </variation>
+        <variation name="InjectableTestCaseVariation2" method="test1">
+            <data name="fixture/data/locator" xsi:type="string">#hplogo</data>
+            <data name="fixture/data/strategy" xsi:type="string">css selector</data>
+            <data name="tag" xsi:type="string">success</data>
+            <constraint name="Magento\Mtf\Test\Constraint\PageOpenSuccess" />
+        </variation>
+    </testCase>
+</config>
