+++ tests/app/Magento/Tag/Test/Handler/Tag/Ui.php
@@ -6,16 +6,21 @@
 
 namespace Magento\Tag\Test\Handler\Tag;
 
-use Magento\Tag\Test\Handler\Tag\TagInterface;
 use Magento\Mtf\Fixture\FixtureInterface;
 use Magento\Mtf\Handler\Ui as AbstractUi;
 
 /**
- * Class Ui
+ * UI handler for Tag fixture.
  */
 class Ui extends AbstractUi implements TagInterface
 {
-   public function persist(FixtureInterface $fixture = null)
+    /**
+     * Persist Tag fixture.
+     *
+     * @param FixtureInterface $fixture [optional]
+     * @return mixed
+     */
+    public function persist(FixtureInterface $fixture = null)
     {
         //
     }
