+++ tests/app/Magento/Tag/Test/TestCase/TagTest.xml
@@ -6,33 +6,33 @@
  */
  -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="../../../../../../etc/variations.xsd">
-  <testCase name="Magento\Tag\Test\TestCase\TagTest">
-    <variation name="TagTestVariation1">
-      <data name="fixture/data/locator" xsi:type="string">#hplogo</data>
-      <data name="fixture/data/strategy" xsi:type="string">css selector</data>
-      <data name="fixture/data/search" xsi:type="string">variation priority:high</data>
-      <data name="tag" xsi:type="string">priority:high</data>
-      <constraint name="Magento\Tag\Test\Constraint\TagSeverityHigh"/>
-      <constraint name="Magento\Tag\Test\Constraint\TagSeverityMiddle"/>
-      <constraint name="Magento\Tag\Test\Constraint\TagSeverityLow"/>
-    </variation>
-    <variation name="TagTestVariation2">
-      <data name="fixture/data/locator" xsi:type="string">#hplogo</data>
-      <data name="fixture/data/strategy" xsi:type="string">css selector</data>
-      <data name="fixture/data/search" xsi:type="string">variation priority:middle</data>
-      <data name="tag" xsi:type="string">priority:middle</data>
-      <constraint name="Magento\Tag\Test\Constraint\TagSeverityHigh"/>
-      <constraint name="Magento\Tag\Test\Constraint\TagSeverityMiddle"/>
-      <constraint name="Magento\Tag\Test\Constraint\TagSeverityLow"/>
-    </variation>
-    <variation name="TagTestVariation3">
-      <data name="fixture/data/locator" xsi:type="string">#hplogo</data>
-      <data name="fixture/data/strategy" xsi:type="string">css selector</data>
-      <data name="fixture/data/search" xsi:type="string">variation priority:low</data>
-      <data name="tag" xsi:type="string">priority:low</data>
-      <constraint name="Magento\Tag\Test\Constraint\TagSeverityHigh"/>
-      <constraint name="Magento\Tag\Test\Constraint\TagSeverityMiddle"/>
-      <constraint name="Magento\Tag\Test\Constraint\TagSeverityLow"/>
-    </variation>
-  </testCase>
+    <testCase name="Magento\Tag\Test\TestCase\TagTest" summary="Check tagging functionality">
+        <variation name="TagTestVariation1">
+            <data name="fixture/data/locator" xsi:type="string">#hplogo</data>
+            <data name="fixture/data/strategy" xsi:type="string">css selector</data>
+            <data name="fixture/data/search" xsi:type="string">variation priority:high</data>
+            <data name="tag" xsi:type="string">priority:high</data>
+            <constraint name="Magento\Tag\Test\Constraint\TagSeverityHigh" />
+            <constraint name="Magento\Tag\Test\Constraint\TagSeverityMiddle" />
+            <constraint name="Magento\Tag\Test\Constraint\TagSeverityLow" />
+        </variation>
+        <variation name="TagTestVariation2">
+            <data name="fixture/data/locator" xsi:type="string">#hplogo</data>
+            <data name="fixture/data/strategy" xsi:type="string">css selector</data>
+            <data name="fixture/data/search" xsi:type="string">variation priority:middle</data>
+            <data name="tag" xsi:type="string">priority:middle</data>
+            <constraint name="Magento\Tag\Test\Constraint\TagSeverityHigh" />
+            <constraint name="Magento\Tag\Test\Constraint\TagSeverityMiddle" />
+            <constraint name="Magento\Tag\Test\Constraint\TagSeverityLow" />
+        </variation>
+        <variation name="TagTestVariation3">
+            <data name="fixture/data/locator" xsi:type="string">#hplogo</data>
+            <data name="fixture/data/strategy" xsi:type="string">css selector</data>
+            <data name="fixture/data/search" xsi:type="string">variation priority:low</data>
+            <data name="tag" xsi:type="string">priority:low</data>
+            <constraint name="Magento\Tag\Test\Constraint\TagSeverityHigh" />
+            <constraint name="Magento\Tag\Test\Constraint\TagSeverityMiddle" />
+            <constraint name="Magento\Tag\Test\Constraint\TagSeverityLow" />
+        </variation>
+    </testCase>
 </config>
