--- tests/app/Test/Handler/BlockRenderMtf/BlockRenderMtfInterface.php
@@ -1,17 +1,0 @@
-<?php
-/**
- * Copyright © 2016 Magento. All rights reserved.
- * See COPYING.txt for license details.
- */
-
-namespace \Test\Handler\BlockRenderMtf;
-
-use Magento\Mtf\Handler\HandlerInterface;
-
-/**
- * Interface BlockRenderMtfInterface
- */
-interface BlockRenderMtfInterface extends HandlerInterface
-{
-   //
-}
