--- tests/app/Test/Handler/Fixture/FixtureInterface.php
@@ -1,17 +1,0 @@
-<?php
-/**
- * Copyright © 2016 Magento. All rights reserved.
- * See COPYING.txt for license details.
- */
-
-namespace \Test\Handler\Fixture;
-
-use Magento\Mtf\Handler\HandlerInterface;
-
-/**
- * Interface FixtureInterface
- */
-interface FixtureInterface extends HandlerInterface
-{
-   //
-}
